feat: editing movement name now works

+++ app/src/main/java/net/martinlundberg/a1repmaxtracker/data/database/dao/MovementDao.kt
@@ -21,5 +21,5 @@ interface MovementDao {
     suspend fun insert(movement: MovementEntity): Long
 
     @Query("DELETE FROM movementEntity WHERE id = :id")
-    suspend fun deleteById(id: Int)
+    suspend fun deleteById(id: Long)
 }
